+++ Code/CrowdControlPlayer.cs
@@ -198,7 +198,6 @@ public sealed class CrowdControlPlayer : ModPlayer
 
     public override void ModifyScreenPosition()
     {
-        // TODO: ModifyScreenPosition is not being called by tModLoader in PlayerLoader.
         ModifyScreenPositionHook?.Invoke();
     }