+++ Code/ID/FeatureID.cs
@@ -12,6 +12,7 @@ public static class FeatureID
     public const int ReduceRespawnTime = 2;
     public const int RemoveTombstone = 3;
     public const int TimedEffectDisplay = 4;
+    public const int FoxMorph = 5;
 
     #endregion
 }