+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed "Clear/Explode/Shuffle Inventory" effects not failing when there are no items.
 - Fixed "Screen Shake" effect not pausing when game is paused.
 
+# [1.1.1] - 2012-03-05
+
+- Minor changes.
+
 # [1.1.0] - 2023-01-30
 
 - Added "Curse" effect; temporarily preventing the streamer from using any items.